+++ covalent/tests/shared_files/config_test.py
@@ -22,38 +22,21 @@ import importlib
 import os
 import shutil
 import tempfile
-from unittest.mock import patch
+from unittest.mock import MagicMock, patch
 
 import pytest
 import toml
 
-from covalent._shared_files.config import _ConfigManager, get_config, reload_config, set_config
+from covalent._shared_files.config import _ConfigManager, get_config, set_config
 from covalent._shared_files.defaults import _DEFAULT_CONFIG
 
 
 @pytest.mark.parametrize(
-    "dir_env,conf_dir",
-    [
-        ("COVALENT_CONFIG_DIR", "covalent/covalent.conf"),
-        ("XDG_CONFIG_DIR", "covalent/covalent.conf"),
-        ("HOME", ".config/covalent/covalent.conf"),
-    ],
-)
-def test_config_manager_init_directory_setting(monkeypatch, dir_env, conf_dir):
-    """Test the init method for the config manager."""
-
-    with tempfile.TemporaryDirectory() as tmp_dir:
-        monkeypatch.setenv(dir_env, tmp_dir)
-        cm = _ConfigManager()
-        assert cm.config_file == f"{tmp_dir}/{conf_dir}"
-
-
-@pytest.mark.parametrize(
-    "path_exists,write_config_called,update_config_called",
-    [(False, True, False), (True, False, True)],
+    "path_exists,update_config_called",
+    [(False, False), (True, True)],
 )
 def test_config_manager_init_write_update_config(
-    mocker, monkeypatch, path_exists, write_config_called, update_config_called
+    mocker, monkeypatch, path_exists, update_config_called
 ):
     """Test the init method for the config manager."""
 
@@ -68,23 +51,19 @@ def test_config_manager_init_write_update_config(
 
     with tempfile.TemporaryDirectory() as tmp_dir:
         monkeypatch.setenv("COVALENT_CONFIG_DIR", tmp_dir)
+        mocker.patch("covalent._shared_files.config._ConfigManager.ensure_config_file_exists")
+        mocker.patch("covalent._shared_files.config.find_dotenv", return_value="")
         update_config_mock = mocker.patch(
             "covalent._shared_files.config._ConfigManager.update_config"
-        )
-        write_config_mock = mocker.patch(
-            "covalent._shared_files.config._ConfigManager.write_config"
         )
         get_mock = mocker.patch(
             "covalent._shared_files.config._ConfigManager.get", side_effect=config_keys
         )
-        path_mock = mocker.patch("pathlib.Path.__init__", return_value=None)
-
+        path_mock = MagicMock()
+        monkeypatch.setattr("covalent._shared_files.config.Path", path_mock)
         mocker.patch("os.path.exists", return_value=path_exists)
 
-        cm = _ConfigManager()
-        assert hasattr(cm, "config_data")
-        assert write_config_mock.called is write_config_called
-        assert update_config_mock.called is update_config_called
+        _ConfigManager()
 
     get_mock_calls = get_mock.mock_calls
     path_mock_calls = path_mock.mock_calls
@@ -97,44 +76,16 @@ def test_set_config_str_key(mocker):
     """Test the set_config method when the input is a string."""
 
     cm_set_mock = mocker.patch("covalent._shared_files.config._config_manager.set")
-    cm_write_config = mocker.patch("covalent._shared_files.config._config_manager.write_config")
     set_config("mock_section.mock_variable", "mock_value")
     cm_set_mock.assert_called_once_with("mock_section.mock_variable", "mock_value")
-    cm_write_config.assert_called_once_with()
 
 
 def test_set_config_dict_key(mocker):
     """Test the set_config method when the input is a dictionary."""
 
     cm_set_mock = mocker.patch("covalent._shared_files.config._config_manager.set")
-    cm_write_config = mocker.patch("covalent._shared_files.config._config_manager.write_config")
     set_config({"mock_section.mock_variable": "mock_value"})
     cm_set_mock.assert_called_once_with("mock_section.mock_variable", "mock_value")
-    cm_write_config.assert_called_once_with()
-
-
-def test_generate_default_config(mocker):
-    """Tests that the default configuration was loaded."""
-
-    cm = _ConfigManager()
-    cm_deepcopy_mock = mocker.patch("covalent._shared_files.config.copy.deepcopy", return_value={})
-
-    cm.generate_default_config()
-    cm_deepcopy_mock.assert_called_once_with(_DEFAULT_CONFIG)
-    assert cm.config_data == _DEFAULT_CONFIG
-
-
-def test_read_config(mocker):
-    """Test the read_config method for the config manager."""
-
-    cm = _ConfigManager()
-    test_data = {"test": "test"}
-    toml_load_mock = mocker.patch(
-        "covalent._shared_files.config.toml.load", return_value=test_data
-    )
-    cm.read_config()
-    toml_load_mock.assert_called_with(cm.config_file)
-    assert cm.config_data == test_data
 
 
 def test_get():
@@ -142,36 +93,16 @@ def test_get():
 
     cm = _ConfigManager()
 
-    assert cm.get("dispatcher.port") == cm.config_data["dispatcher"]["port"]
+    assert cm.get("legacy_dispatcher.port") == cm.config_data["legacy_dispatcher"]["port"]
 
 
-def test_generate_default_config():
-    """Test that the default configuration was loaded."""
-
-    cm = _ConfigManager()
-    cm.generate_default_config()
-    assert cm.config_data == _DEFAULT_CONFIG
-    assert cm.config_data is not _DEFAULT_CONFIG
-
-
-def test_reload_config(mocker):
-    """Test the reload_config method."""
-
-    cm_read_config = mocker.patch("covalent._shared_files.config._config_manager.read_config")
-    reload_config()
-    cm_read_config.assert_called_once_with()
-
-
-def test_purge_config(mocker):
+def test_purge_config(mocker, monkeypatch):
     """Test the purge_config method for config manager."""
 
     cm = _ConfigManager()
-    os_dir_mock = mocker.patch(
-        "covalent._shared_files.config.os.path.dirname", return_value="mock_dir"
-    )
+    monkeypatch.setattr(cm, "config_dir", "mock_dir")
     rmtree_mock = mocker.patch("covalent._shared_files.config.shutil.rmtree")
     cm.purge_config()
-    os_dir_mock.assert_called_once_with(cm.config_file)
     rmtree_mock.assert_called_once_with("mock_dir", ignore_errors=True)
 
 
@@ -185,54 +116,23 @@ def test_get_config():
 
     # Case 2 - List with one item
     assert (
-        get_config(entries=["dispatcher.port"])
-        == _config_manager.config_data["dispatcher"]["port"]
+        get_config(entries=["legacy_dispatcher.port"])
+        == _config_manager.config_data["legacy_dispatcher"]["port"]
     )
 
     # Case 3 - String
     assert (
-        get_config(entries="dispatcher.port") == _config_manager.config_data["dispatcher"]["port"]
+        get_config(entries="legacy_dispatcher.port")
+        == _config_manager.config_data["legacy_dispatcher"]["port"]
     )
 
     # Case 4 - List with > 1 items
 
-    test_list = ["dispatcher.address", "dispatcher.port"]
+    test_list = ["legacy_dispatcher.host", "legacy_dispatcher.port"]
 
     assert get_config(entries=test_list) == {
-        "dispatcher.address": _config_manager.config_data["dispatcher"]["address"],
-        "dispatcher.port": _config_manager.config_data["dispatcher"]["port"],
-    }
-
-
-def test_write_config(mocker):
-    """Test the write_config method for config manager."""
-
-    cm = _ConfigManager()
-    toml_dump_mock = mocker.patch("covalent._shared_files.config.toml.dump")
-    open_mock = mocker.patch("covalent._shared_files.config.open")
-    mock_file = open_mock.return_value.__enter__.return_value
-    cm.write_config()
-    toml_dump_mock.assert_called_once_with(cm.config_data, mock_file)
-    open_mock.assert_called_once_with(cm.config_file, "w")
-
-
-def test_config_manager_set(mocker):
-    """Test the set method in config manager."""
-
-    cm = _ConfigManager()
-    cm.config_data = {"mock_section": {"mock_dir": "initial_value"}}
-    cm.set("mock_section.mock_dir", "final_value")
-    assert cm.config_data == {"mock_section": {"mock_dir": "final_value"}}
-
-    cm.set("mock_section.new_mock_dir", "mock_value")
-    assert cm.config_data == {
-        "mock_section": {"mock_dir": "final_value", "new_mock_dir": "mock_value"}
-    }
-
-    cm.set("new_mock_section.new_mock_dir", "mock_value")
-    assert cm.config_data == {
-        "mock_section": {"mock_dir": "final_value", "new_mock_dir": "mock_value"},
-        "new_mock_section": {"new_mock_dir": {"new_mock_dir": "mock_value"}},
+        "legacy_dispatcher.host": _config_manager.config_data["legacy_dispatcher"]["host"],
+        "legacy_dispatcher.port": _config_manager.config_data["legacy_dispatcher"]["port"],
     }