+++ refactor/runner/app/core/config.py
@@ -21,7 +21,10 @@
 import os
 from typing import List, Union
 
+from dotenv import find_dotenv, load_dotenv
 from pydantic import AnyHttpUrl, BaseSettings, validator
+
+load_dotenv()
 
 
 class Settings(BaseSettings):
@@ -44,6 +47,10 @@ class Settings(BaseSettings):
 
     class Config:
         case_sensitive = True
+        env_file = ".env"
 
 
 settings = Settings()
+
+print("Configuration:")
+print(settings)