+++ CHANGELOG.md
@@ -12,6 +12,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Installation using `pip install -e .` is fixed with regards to the nats installation.
 - Several missing `__init__.py` files are now included.
 
+## [0.79.0] - 2022-04-14
+
+### Added
+
+- Covalent `config` cli command to alter config values or display covalent configuration
+
+### Changed
+
+- Removed environment section from Supervisord config in order to read from root `.env` file instead
+- Refactored config manager to use project root `.env` file for configuration
+
 ## [0.78.0] - 2022-04-13
 
 ### Changed